% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bartext.R
\name{geom_bartext}
\alias{geom_bartext}
\title{Repelling text for GeomBar.}
\usage{
geom_bartext(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", parse = FALSE, nudge_x = 0, nudge_y = 0,
  spacing = 0.005, dir = "v", check_overlap = FALSE, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)
}
\description{
Repelling text for GeomBar.
}
