% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_share.R, R/geom_arcbar.R,
%   R/geom_boxjitter.R, R/geom_circle.R, R/geom_parliament.R,
%   R/geom_tshighlight.R, R/ggproto-classes.R, R/stat-arcbar.R,
%   R/stat-boxjitter.R, R/stat-circle.R, R/stat-parliament.R
\docType{data}
\name{FacetShare}
\alias{FacetShare}
\alias{GeomArcBar}
\alias{GeomBoxJitter}
\alias{GeomCircle}
\alias{GeomParliament}
\alias{GeomTshighlight}
\alias{ggpol-extensions}
\alias{StatArcBar}
\alias{StatBoxJitter}
\alias{StatCircle}
\alias{StatParliament}
\title{ggplot extensions}
\description{
ggplot extensions
}
\keyword{datasets}
