% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bartext.R
\name{geom_bartext}
\alias{geom_bartext}
\title{Repelling text for GeomBar.}
\usage{
geom_bartext(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", parse = FALSE, nudge_x = 0, nudge_y = 0,
  spacing = 0.003, dir = "v", check_overlap = FALSE, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{Use to override the default connection between
\code{geom_boxplot} and \code{stat_boxplot}.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{spacing}{Defaults to 0.003. Minimum spacing between labels in NPC units.}

\item{dir}{Defaults to "v", i.e. vertical repel of overlapping groups of labels. Can alternatively be set to "h" for horizontal repel.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
Repelling text for GeomBar.
}
\examples{
df <- data.frame(L = rep(LETTERS[1:2], each = 4),
                 l = rep(letters[1:4], 2),
                 val = c(96.5, 1, 2, 0.5, 48, 0.7, 0.3, 51))
 
ggplot(df, aes(x = L, y = val, fill = l)) +
  geom_bar(stat = "identity") +
  geom_bartext(aes(label = scales::percent(val / 100)), position = position_stack(vjust = 0.5)) +
  ggtitle("GeomBartext") 
  
}
