% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arcbar.R, R/geom_bartext.R,
%   R/geom_boxjitter.R, R/geom_circle.R, R/geom_parliament.R,
%   R/geom_tshighlight.R, R/ggproto-classes.R, R/stat-arcbar.R,
%   R/stat-boxjitter.R, R/stat-circle.R, R/stat-confmat.R, R/stat-parliament.R
\docType{data}
\name{GeomArcbar}
\alias{GeomArcbar}
\alias{GeomBartext}
\alias{GeomBoxJitter}
\alias{GeomCircle}
\alias{GeomParliament}
\alias{GeomTshighlight}
\alias{ggpol-extensions}
\alias{StatArcbar}
\alias{StatBoxJitter}
\alias{StatCircle}
\alias{StatConfmat}
\alias{StatParliament}
\title{ggplot extensions}
\description{
ggplot extensions
}
\keyword{datasets}
